Use errors instead of panics
The anyhow crate helps a lot in error management. Its Result can be built from any error type
defined in a library, and also adding context information when errors are propagated.

A rule of thumb for error management:
- libraries should define their own semantically-rich error types. The thiserror crate helps
  implementing these errors
- applications can use anyhow to easily handle a variety of errors coming from libraries.

This commit also adds a "try_map" extension to iterators to make it easier write a chain or
faillible (i.e. can return errors) transformation in a functional way.